feat: move the get triples per field into a utils method
Refs: #SFB-8
